Task 2/teeth_app/requirements.txt
# Pinned dependencies for reproducible builds
streamlit==1.33.0
tensorflow==2.16.1
numpy==1.26.4
pillow==10.2.0
typing_extensions==4.10.0
